requirements.txt
google-api-python-client==2.136.0
youtube-transcript-api==0.6.2
google-generativeai==0.7.2
requests==2.32.3
